+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased — 0.5 API transition
+## 0.5.0 — 2026-07-27
 
 ### Deprecated